Modify: database path and Mongoose env type
