+++ back/src/main.ts
@@ -1,6 +1,6 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
-import { useEnv } from './settings/environment';
+import { useEnv } from './utils/environment';
 import { winstonLogger } from '@src/utils/winston';
 
 useEnv();
